Quick bug fix that will hopefully take care of UI issues.

+++ cat-web/src/main/java/com/digitalbarista/cat/ejb/session/FacebookManager.java
@@ -258,7 +258,7 @@ public class FacebookManager
 		crit.add(Restrictions.eq("facebookAppName", app.getAppName()));
 		crit.setProjection(Projections.rowCount());
 
-		Integer count = (Integer) crit.uniqueResult();
+		Integer count = ((Long) crit.uniqueResult()).intValue();
 
 		updateMessageCounter(appName, uid, count);
 	}
